Results, with `-O2`:

Without inline pragmas:
```
benchmarking clique/5  
time                 91.99 μs   (91.66 μs .. 92.30 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 91.56 μs   (91.38 μs .. 91.79 μs)
std dev              726.6 ns   (591.5 ns .. 871.6 ns)
                       
benchmarking clique/10 
time                 322.1 μs   (321.7 μs .. 322.5 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 322.4 μs   (322.1 μs .. 322.8 μs)
std dev              1.334 μs   (1.139 μs .. 1.592 μs)
                       
benchmarking clique/25 
time                 2.162 ms   (2.153 ms .. 2.170 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.168 ms   (2.163 ms .. 2.173 ms)
std dev              17.29 μs   (14.60 μs .. 20.45 μs)
                       
benchmarking clique/40 
time                 5.521 ms   (5.506 ms .. 5.537 ms)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 5.555 ms   (5.541 ms .. 5.588 ms)
std dev              62.89 μs   (38.44 μs .. 109.5 μs)
                       
benchmarking line/5    
time                 88.36 μs   (88.26 μs .. 88.47 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 88.42 μs   (88.32 μs .. 88.52 μs)
std dev              311.5 ns   (263.7 ns .. 384.9 ns)
                       
benchmarking line/10   
time                 368.5 μs   (367.7 μs .. 369.8 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 368.2 μs   (367.8 μs .. 369.0 μs)
std dev              1.885 μs   (1.374 μs .. 2.732 μs)
                       
benchmarking line/25   
time                 3.145 ms   (3.137 ms .. 3.153 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 3.142 ms   (3.137 ms .. 3.148 ms)
std dev              17.62 μs   (14.14 μs .. 22.19 μs)
                       
benchmarking line/40   
time                 10.54 ms   (10.53 ms .. 10.56 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 10.53 ms   (10.52 ms .. 10.54 ms)
std dev              36.28 μs   (27.38 μs .. 52.41 μs)
                       
benchmarking loop/50   
time                 472.9 μs   (471.4 μs .. 475.2 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 472.6 μs   (471.2 μs .. 474.5 μs)
std dev              5.409 μs   (4.134 μs .. 7.072 μs)
                       
benchmarking loop/100  
time                 1.393 ms   (1.391 ms .. 1.396 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.395 ms   (1.392 ms .. 1.400 ms)
std dev              10.73 μs   (6.322 μs .. 20.65 μs)
                       
benchmarking loop/500  
time                 21.42 ms   (21.26 ms .. 21.62 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 21.30 ms   (21.23 ms .. 21.40 ms)
std dev              199.5 μs   (145.2 μs .. 267.4 μs)
                       
benchmarking loop/1000 
time                 88.64 ms   (86.19 ms .. 91.45 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 85.85 ms   (84.82 ms .. 87.15 ms)
std dev              1.919 ms   (1.247 ms .. 3.101 ms)
                       
benchmarking tight/50  
time                 201.2 μs   (200.8 μs .. 201.7 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 201.9 μs   (201.4 μs .. 202.5 μs)
std dev              1.707 μs   (1.374 μs .. 2.218 μs)
                       
benchmarking tight/100 
time                 732.6 μs   (730.1 μs .. 736.7 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 734.6 μs   (732.5 μs .. 737.3 μs)
std dev              7.665 μs   (6.177 μs .. 9.657 μs)
                       
benchmarking tight/500 
time                 16.79 ms   (16.72 ms .. 16.96 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 16.84 ms   (16.78 ms .. 17.05 ms)
std dev              248.7 μs   (80.32 μs .. 484.4 μs)
                       
benchmarking tight/1000
time                 72.76 ms   (72.10 ms .. 73.43 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 72.15 ms   (71.94 ms .. 72.41 ms)
std dev              415.2 μs   (296.4 μs .. 563.1 μs)
                       
benchmarking parse/50  
time                 6.102 ms   (6.030 ms .. 6.164 ms)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 6.043 ms   (6.018 ms .. 6.086 ms)
std dev              91.41 μs   (68.09 μs .. 144.4 μs)
                       
benchmarking parse/100 
time                 12.01 ms   (11.80 ms .. 12.28 ms)
                     0.999 R²   (0.997 R² .. 1.000 R²)
mean                 11.93 ms   (11.86 ms .. 12.01 ms)
std dev              202.7 μs   (156.5 μs .. 286.5 μs)
                       
benchmarking parse/500 
time                 66.70 ms   (66.28 ms .. 67.16 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 66.82 ms   (66.64 ms .. 67.05 ms)
std dev              350.8 μs   (249.0 μs .. 502.3 μs)
                       
benchmarking parse/1000
time                 135.9 ms   (133.5 ms .. 139.1 ms)
                     0.999 R²   (0.998 R² .. 1.000 R²)
mean                 133.8 ms   (133.0 ms .. 135.5 ms)
std dev              1.664 ms   (648.6 μs .. 2.637 ms)
variance introduced by outliers: 11% (moderately inflated)
```

With inline pragmas:
```
benchmarking clique/5  
time                 40.86 μs   (40.69 μs .. 41.08 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 41.67 μs   (41.30 μs .. 42.39 μs)
std dev              1.599 μs   (973.5 ns .. 2.822 μs)
variance introduced by outliers: 42% (moderately inflated)
                       
benchmarking clique/10 
time                 157.8 μs   (140.8 μs .. 181.6 μs)
                     0.942 R²   (0.910 R² .. 0.996 R²)
mean                 156.0 μs   (149.9 μs .. 165.2 μs)
std dev              26.43 μs   (17.68 μs .. 38.85 μs)
variance introduced by outliers: 92% (severely inflated)
                       
benchmarking clique/25 
time                 1.097 ms   (1.042 ms .. 1.172 ms)
                     0.985 R²   (0.978 R² .. 0.998 R²)
mean                 1.082 ms   (1.063 ms .. 1.116 ms)
std dev              81.99 μs   (56.92 μs .. 109.5 μs)
variance introduced by outliers: 59% (severely inflated)
                       
benchmarking clique/40 
time                 2.699 ms   (2.669 ms .. 2.735 ms)
                     0.997 R²   (0.992 R² .. 0.999 R²)
mean                 2.776 ms   (2.737 ms .. 2.863 ms)
std dev              181.1 μs   (98.46 μs .. 321.3 μs)
variance introduced by outliers: 45% (moderately inflated)
                       
benchmarking line/5    
time                 42.00 μs   (41.17 μs .. 43.07 μs)
                     0.997 R²   (0.992 R² .. 1.000 R²)
mean                 41.22 μs   (40.87 μs .. 42.20 μs)
std dev              1.780 μs   (907.4 ns .. 3.697 μs)
variance introduced by outliers: 48% (moderately inflated)
                       
benchmarking line/10   
time                 150.9 μs   (149.2 μs .. 154.2 μs)
                     0.994 R²   (0.984 R² .. 0.999 R²)
mean                 157.0 μs   (153.5 μs .. 163.3 μs)
std dev              14.78 μs   (9.891 μs .. 22.11 μs)
variance introduced by outliers: 79% (severely inflated)
                       
benchmarking line/25   
time                 1.018 ms   (1.012 ms .. 1.026 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 1.011 ms   (1.009 ms .. 1.015 ms)
std dev              10.85 μs   (8.285 μs .. 14.13 μs)
                       
benchmarking line/40   
time                 2.832 ms   (2.796 ms .. 2.855 ms)
                     0.998 R²   (0.994 R² .. 1.000 R²)
mean                 2.948 ms   (2.891 ms .. 3.131 ms)
std dev              312.3 μs   (62.31 μs .. 627.6 μs)
variance introduced by outliers: 69% (severely inflated)
                       
benchmarking loop/50   
time                 211.5 μs   (211.2 μs .. 211.9 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 211.3 μs   (211.1 μs .. 211.6 μs)
std dev              891.2 ns   (699.0 ns .. 1.166 μs)
                       
benchmarking loop/100  
time                 535.8 μs   (534.4 μs .. 537.8 μs)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 536.3 μs   (534.9 μs .. 540.6 μs)
std dev              7.458 μs   (2.929 μs .. 15.81 μs)
                       
benchmarking loop/500  
time                 5.793 ms   (5.755 ms .. 5.846 ms)
                     0.994 R²   (0.978 R² .. 1.000 R²)
mean                 5.877 ms   (5.821 ms .. 6.073 ms)
std dev              294.3 μs   (73.69 μs .. 601.1 μs)
variance introduced by outliers: 28% (moderately inflated)
                       
benchmarking loop/1000 
time                 19.75 ms   (19.34 ms .. 20.11 ms)
                     0.999 R²   (0.997 R² .. 0.999 R²)
mean                 19.40 ms   (19.26 ms .. 19.57 ms)
std dev              355.0 μs   (283.4 μs .. 441.6 μs)
                       
benchmarking tight/50  
time                 56.52 μs   (56.27 μs .. 56.84 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 56.31 μs   (56.13 μs .. 56.47 μs)
std dev              565.7 ns   (432.9 ns .. 733.5 ns)
                       
benchmarking tight/100 
time                 159.9 μs   (159.1 μs .. 161.1 μs)
                     0.999 R²   (0.999 R² .. 1.000 R²)
mean                 162.4 μs   (161.0 μs .. 166.1 μs)
std dev              7.322 μs   (3.956 μs .. 12.40 μs)
variance introduced by outliers: 44% (moderately inflated)
                       
benchmarking tight/500 
time                 3.141 ms   (3.011 ms .. 3.295 ms)
                     0.983 R²   (0.972 R² .. 0.992 R²)
mean                 3.005 ms   (2.940 ms .. 3.107 ms)
std dev              266.3 μs   (184.3 μs .. 413.3 μs)
variance introduced by outliers: 60% (severely inflated)
                       
benchmarking tight/1000
time                 11.11 ms   (10.82 ms .. 11.50 ms)
                     0.993 R²   (0.985 R² .. 1.000 R²)
mean                 11.26 ms   (11.11 ms .. 11.52 ms)
std dev              490.8 μs   (293.8 μs .. 751.2 μs)
variance introduced by outliers: 19% (moderately inflated)
                       
benchmarking parse/50  
time                 6.340 ms   (6.194 ms .. 6.516 ms)
                     0.991 R²   (0.976 R² .. 0.999 R²)
mean                 6.220 ms   (6.152 ms .. 6.413 ms)
std dev              304.4 μs   (143.4 μs .. 596.7 μs)
variance introduced by outliers: 26% (moderately inflated)
                       
benchmarking parse/100 
time                 12.12 ms   (11.60 ms .. 12.56 ms)
                     0.989 R²   (0.976 R² .. 0.998 R²)
mean                 12.22 ms   (11.96 ms .. 12.78 ms)
std dev              937.4 μs   (435.5 μs .. 1.487 ms)
variance introduced by outliers: 39% (moderately inflated)
                       
benchmarking parse/500 
time                 68.21 ms   (66.76 ms .. 71.27 ms)
                     0.998 R²   (0.994 R² .. 1.000 R²)
mean                 68.01 ms   (67.35 ms .. 68.92 ms)
std dev              1.438 ms   (886.0 μs .. 2.083 ms)
                       
benchmarking parse/1000
time                 134.5 ms   (133.2 ms .. 135.8 ms)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 133.3 ms   (132.6 ms .. 133.9 ms)
std dev              967.8 μs   (595.8 μs .. 1.503 ms)
variance introduced by outliers: 11% (moderately inflated)
```